contractflow/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0
pydantic>=2.10,<3.0
pydantic-settings>=2.6,<3.0
email-validator>=2.2,<3.0
python-multipart>=0.0.20,<1.0
argon2-cffi>=23.1,<24.0
python-magic>=0.4.27,<1.0
python-jose[cryptography]>=3.3,<4.0
httpx>=0.28,<1.0
redis>=5.2,<6.0
azure-storage-blob>=12.24,<13.0
azure-identity>=1.19,<2.0
azure-keyvault-secrets>=4.9,<5.0
opentelemetry-api>=1.29,<2.0
opentelemetry-sdk>=1.29,<2.0
opentelemetry-instrumentation-fastapi>=0.50b,<1.0
